Extract generic stream decoder adaptor
