+++ src/stream/generic/mod.rs
@@ -1,3 +1,4 @@
+mod decoder;
 mod encoder;
 
-pub(crate) use encoder::Encoder;
+pub(crate) use self::{decoder::Decoder, encoder::Encoder};
